Refactor Appointment and Scheduler classes to use NotifyScheduler for observer notifications
